(feat) Outline tech support convoyer control outline working with several camera
