--- ConcourUbisoft/Assets/Scripts/TechSupport/Controller/ControllerSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a76f036a8366ac4bb0ccadc7555ed66
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
