+++ ConcourUbisoft/Assets/Scripts/TechSupport/GridSystem.cs
@@ -1,8 +1,7 @@
 using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
-namespace TechSupport.Surveillance
+namespace TechSupport
 {
     public class GridSystem
     {
