--- ConcourUbisoft/Assets/Scripts/TechSupport/Surveillance/SurveillanceSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02ed7229643a50041beb3674a435da53
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
