+++ ConcourUbisoft/ProjectSettings/InputManager.asset
@@ -755,7 +755,7 @@ InputManager:
     invert: 0
     type: 0
     axis: 9
-    joyNum: 8
+    joyNum: 0
   - serializedVersion: 3
     m_Name: ControlXBO
     descriptiveName: b
@@ -770,8 +770,8 @@ InputManager:
     snap: 0
     invert: 0
     type: 0
-    axis: 9
-    joyNum: 8
+    axis: 0
+    joyNum: 0
   - serializedVersion: 3
     m_Name: ControlPS
     descriptiveName: y
@@ -786,5 +786,5 @@ InputManager:
     snap: 0
     invert: 0
     type: 0
-    axis: 9
-    joyNum: 8
+    axis: 0
+    joyNum: 0
